test: simplify code

+++ src/dense/iter.rs
@@ -862,7 +862,7 @@ mod tests {
             matrix
                 .iter_elements_mut_with_index()
                 .for_each(|(index, element)| {
-                    *element += index.row as i32 + index.col as i32;
+                    *element += index.row + index.col;
                 });
             let expected = matrix![[1, 3, 5], [5, 7, 9]];
             assert_eq!(matrix, expected);
